Extract Uniswap artefacts to lib
