+++ test/uniswapManager_invest.js
@@ -1,8 +1,8 @@
 const etherlime = require('etherlime-lib');
 const Wallet = require("../build/BaseWallet");
 const ModuleRegistry = require("../build/ModuleRegistry");
-const UniswapFactory = require("../contracts/test/uniswap/UniswapFactory");
-const UniswapExchange = require("../contracts/test/uniswap/UniswapExchange");
+const UniswapFactory = require("../lib/uniswap/UniswapFactory");
+const UniswapExchange = require("../lib/uniswap/UniswapExchange");
 const UniswapManager = require("../build/UniswapManager");
 const GuardianStorage = require("../build/GuardianStorage");
 const ERC20 = require("../build/TestERC20");
